Move QArrayLike type definition to inside QArray to avoid circular import
