+++ dynamiqs/qarrays/qarray.py
@@ -3,21 +3,27 @@ from __future__ import annotations
 import logging
 from abc import abstractmethod
 from math import prod
-from typing import TYPE_CHECKING
+from typing import Any, Union, get_args
 
 import equinox as eqx
 import jax.numpy as jnp
 import numpy as np
 from equinox.internal._omega import _Metaω  # noqa: PLC2403
 from jax import Array, Device
+from jaxtyping import ArrayLike
 from qutip import Qobj
 
 from .layout import Layout
 
-if TYPE_CHECKING:  # avoid circular import by importing only during type checking
-    from .types import QArrayLike
+__all__ = ['QArray', 'QArrayLike', 'isqarraylike']
 
-__all__ = ['QArray']
+
+def isqarraylike(x: Any) -> bool:
+    if isinstance(x, get_args(_QArrayLike)):
+        return True
+    elif isinstance(x, list):
+        return all(isqarraylike(_x) for _x in x)
+    return False
 
 
 class QArray(eqx.Module):
@@ -451,3 +457,21 @@ def _include_last_two_dims(axis: int | tuple[int, ...] | None, ndim: int) -> boo
     return axis is None or (
         ndim - 1 in [a % ndim for a in axis] and ndim - 2 in [a % ndim for a in axis]
     )
+
+# In this file we define an extended array type named `QArrayLike`. Most
+# functions in the library take a `QArrayLike` as argument and return a `QArray`.
+# `QArrayLike` can be:
+# - any numeric type (bool, int, float, complex),
+# - a JAX array,
+# - a NumPy array,
+# - a QuTiP Qobj,
+# - a dynamiqs QArray,
+# - a nested list of these types.
+# An object of type `QArrayLike` can be converted to a `QArray` with `asqarray`.
+
+# extended array like type
+_QArrayLike = Union[ArrayLike, QArray, Qobj]
+# a type alias for nested list of _QArrayLike
+_NestedQArrayLikeList = list[Union[_QArrayLike, '_NestedQArrayLikeList']]
+# a type alias for any type compatible with asqarray
+QArrayLike = Union[_QArrayLike, _NestedQArrayLikeList]
